+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 ## unreleased
+- [[#22] make log level configurable and log as json](https://github.com/leberKleber/simple-jwt-provider/issues/22)
+
 ## v1.1.0
 - [[#3] secure admin api password config](https://github.com/leberKleber/simple-jwt-provider/issues/10)
 - [[#9] make jwt lifetime configurable](https://github.com/leberKleber/simple-jwt-provider/issues/9)